Go back to a block-style cursor.
I really like bar-style cursors, but there are some things which Emacs
doesn't do too well with them. The most important is that Emacs, when
you click, treats the *glyph* as the thing you clicked and not
the *space* in between the glyph. This makes it hard to use a
bar-style cursor since it assumes you are clicking in between. So in
turn you select the wrong character. Argh!

Also there is some visually hard things to get into grips with when
considering the bar-style cursor. It is harder to see, so you almost
need blinking enabled, but I don't particularly like blinking cursors
since they distract me, when I am typing.

So kill the blinking, but take the block-cursor.