Try using multiple-cursors a bit and see
Major change: selected changes the way the region works, by making it
not delete stuff if you start typing, but you have to type C-d to
delete the region. I think this is a good trade-off if you are getting
a selected-keymap in return and can do stuff on selected text.